requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

Django == 2.0.3
Pillow
djangorestframework
gunicorn == 19.7.0
psycopg2
whitenoise
social-auth-app-django